backend/requirements/base.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis & Caching
redis==5.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Processing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Excel & PDF Export
openpyxl==3.1.2
reportlab==4.0.9

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2